+++ CHANGELOG.rst
@@ -53,6 +53,10 @@
     **breaking change** as somebody's code could depend on this
     IndexError. At this stage it seems an acceptable risk.
 
+- Improvements
+
+  - Fix several typing issues with SpectrumNDCollection.
+
 - Other changes
 
   - Error messages have been overhauled and now follow a consistent format::